♻ Create component Scrollable from Chat page for future reuse
