+++ src/pages/Chat.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef, useState } from "react";
+import React, { useState } from "react";
 import styled from "styled-components";
 
 import Form from "react-bootstrap/Form";
@@ -8,6 +8,7 @@ import { faPaperPlane } from "@fortawesome/free-solid-svg-icons";
 
 import JobHeader from "../components/headers/JobHeader";
 import useMessages from "../hooks/useMessages";
+import Scrollable from "../components/Scrollable";
 
 const MessageContainer = styled.div`
   width: 100%;
@@ -49,73 +50,6 @@ function Message({ children: text, isFromOwner = false }) {
         </div>
       </MessageInnerContainer>
     </MessageContainer>
-  );
-}
-
-const MessagesContentContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  flex-grow: 1;
-
-  height: 320px;
-  overflow-y: scroll;
-
-  box-shadow: ${(props) =>
-    props.showShadow
-      ? "inset 0px -131px 100px -140px rgba(0, 0, 0, 0.2)"
-      : undefined};
-
-  // Hide scrollbar
-  ::-webkit-scrollbar {
-    display: none; /* Safari and Chrome */
-  }
-  scrollbar-width: none; /* Firefox */
-`;
-
-function isScrollable(element) {
-  return (
-    element.scrollWidth > element.clientWidth ||
-    element.scrollHeight > element.clientHeight
-  );
-}
-
-function MessagesContent({ messages }) {
-  const containerRef = useRef(null);
-
-  const [showShadow, setShowShadow] = useState(true);
-
-  const [canScroll, setCanScroll] = useState(false);
-
-  useEffect(() => {
-    setCanScroll(isScrollable(containerRef.current));
-  }, [messages]);
-
-  const handleScroll = (e) => {
-    const bottom =
-      e.target.scrollHeight - e.target.scrollTop === e.target.clientHeight;
-    if (bottom) {
-      setShowShadow(false);
-    } else {
-      setShowShadow(true);
-    }
-  };
-
-  return (
-    <MessagesContentContainer
-      onScroll={handleScroll}
-      showShadow={canScroll && showShadow}
-      ref={containerRef}
-    >
-      {messages.map((message, index) => (
-        <Message
-          key={`${index} ${message.content}`}
-          isFromOwner={message.isFromOwner}
-        >
-          {message.content}
-        </Message>
-      ))}
-    </MessagesContentContainer>
   );
 }
 
@@ -183,7 +117,17 @@ function Messages() {
 
   return (
     <MessagesContainer>
-      <MessagesContent messages={messages} />
+      <Scrollable>
+        {messages.map((message, index) => (
+          <Message
+            key={`${index} ${message.content}`}
+            isFromOwner={message.isFromOwner}
+          >
+            {message.content}
+          </Message>
+        ))}
+      </Scrollable>
+
       <MessagesInput onSendMessage={(message) => send(message)} />
     </MessagesContainer>
   );
